AddRef in toNative, not toJava to conform to COM
If we AddRef in toJava we double AddRef for out parameters
leading to a memory leak.  If we AddRef in toNative, this
AddRef's the pointer before we hand it out, and then when we pick
it back up in toJava we will either obtain ownership of the ref
we got via toNative->AddRef, or the ref the callee added for an
out parameter.